feat: createToaster util
