+++ ui/applets/kit/src/components/index.ts
@@ -34,6 +34,7 @@ export { AddressVisualizer } from "./AddressVisualizer";
 export { Range } from "./Range";
 export { Tooltip } from "./Tooltip";
 export { Radio } from "./RadioGroup";
+export { Toast } from "./Toast";
 
 /* ---------------------------------- Icons --------------------------------- */
