+++ ui/applets/kit/src/utils/index.ts
@@ -4,3 +4,4 @@ export { forwardRefPolymorphic } from "./polymorph";
 export { ensureErrorMessage } from "./error";
 export { createContext } from "./context";
 export { numberMask } from "./masks";
+export { type ToastController, createToaster } from "./toaster";
