+++ ui/portal/web/src/app.tsx
@@ -1,13 +1,15 @@
-import type React from "react";
-import { Toaster } from "react-hot-toast";
+import { Toast } from "@left-curve/applets-kit";
 import { RootModal } from "./components/modals/RootModal";
 
+import { createToaster } from "@left-curve/applets-kit";
 import { DangoStoreProvider } from "@left-curve/store";
 import { QueryCache, QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { config } from "~/store";
 
 import { AppProvider } from "./app.provider";
 import { AppRouter } from "./app.router";
+
+import type React from "react";
 
 import "../public/global.css";
 import "@left-curve/ui-config/fonts/ABCDiatypeRounded/normal/800.css";
@@ -20,7 +22,8 @@ import "@left-curve/ui-config/fonts/ABCDiatypeRounded/mono/500.css";
 
 import "@left-curve/ui-config/fonts/Exposure/italic/400.css";
 import "@left-curve/ui-config/fonts/Exposure/italic/700.css";
-import { toast } from "./components/foundation/Toast";
+
+const [Toaster, toast] = createToaster((props) => <Toast {...props} />);
 
 const queryClient = new QueryClient({
   queryCache: new QueryCache({
@@ -42,7 +45,7 @@ export const App: React.FC = () => {
   return (
     <DangoStoreProvider config={config}>
       <QueryClientProvider client={queryClient}>
-        <AppProvider>
+        <AppProvider toast={toast}>
           <AppRouter />
           <Toaster position="bottom-center" containerStyle={{ zIndex: 99999999 }} />
           <RootModal />
